[LLT-6204] Extract conntracker to a separate mod
