+++ crates/telio-firewall/src/lib.rs
@@ -4,4 +4,5 @@
 //! Implements stateful firewall to keep track of
 //! initiated connections, and deny inbound packet
 //! from an unrecognized source
+pub(crate) mod conntrack;
 pub mod firewall;
